+++ src/abi/struct.ts
@@ -226,14 +226,14 @@ export const StructFunctions: Map<string, AbiFunction> = new Map([
                 const arg1 = args[1]!;
                 if (arg0.kind !== "ref") {
                     throwCompilationError(
-                        "fromSlice() is implemented only for struct types",
+                        "fromSlice() is implemented only for struct/contract types",
                         ref,
                     );
                 }
                 const tp = getType(ctx, arg0.name);
                 if (tp.kind !== "struct") {
                     throwCompilationError(
-                        "fromSlice() is implemented only for struct types",
+                        "fromSlice() is implemented only for struct/contract types",
                         ref,
                     );
                 }
@@ -256,7 +256,7 @@ export const StructFunctions: Map<string, AbiFunction> = new Map([
                 const arg1 = args[1]!;
                 if (arg0.kind !== "ref") {
                     throwCompilationError(
-                        "fromSlice() is implemented only for struct types",
+                        "fromSlice() is implemented only for struct/contract types",
                         ref,
                     );
                 }