+++ dev-docs/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Language features
+
+- Support `fromCell()` and `fromSlice()` methods on contract types: PR [#3303](https://github.com/tact-lang/tact/pull/3303)
+
+### Release contributors
+
+- [Anton Trunov](https://github.com/anton-trunov)
+
 ## [1.6.12] - 2025-05-27
 
 ### Language features